+++ terraform/vsphere-virtual-machine/content-library-template-linux-guest-customization-hcp-packer/versions.tf
@@ -6,12 +6,12 @@ terraform {
   required_providers {
     hcp = {
       source  = "hashicorp/hcp"
-      version = "~> 0.84.1"
+      version = "~> 0.89.0"
     }
     vsphere = {
       source  = "hashicorp/vsphere"
-      version = ">= 2.7.0"
+      version = ">= 2.8.1"
     }
   }
-  required_version = ">= 1.7.1"
+  required_version = ">= 1.8.3"
 }