+++ terraform/vsphere-virtual-machine/template-linux-guest-customization/versions.tf
@@ -6,8 +6,8 @@ terraform {
   required_providers {
     vsphere = {
       source  = "hashicorp/vsphere"
-      version = ">= 2.7.0"
+      version = ">= 2.8.1"
     }
   }
-  required_version = ">= 1.7.1"
+  required_version = ">= 1.8.3"
 }